Enum versions::Versioning

pub enum Versioning {
    Ideal(SemVer),
    General(Version),
    Complex(Mess),
}

A top-level Versioning type which acts as a wrapper for the more specific types.

Examples

use versions::Versioning;

let a = Versioning::new("1.2.3-1").unwrap();   // SemVer.
let b = Versioning::new("1.2.3r1").unwrap();   // Not SemVer but good enough.
let c = Versioning::new("000.007-1").unwrap(); // Garbage.

assert!(a.is_ideal());
assert!(b.is_general());
assert!(c.is_complex());

Variants

Ideal(SemVer)

Follows good parsing and comparison rules.

General(Version)

A little more permissive than SemVer.

Complex(Mess)

Hope that you need not venture here.

Implementations

impl Versioning

pub fn new(s: &str) -> Option<Versioning>

Create a Versioning by attempting to parse the input first as SemVer, then as a Version, and finally as a Mess.

pub fn parse(i: &str) -> IResult<&str, Versioning>

The raw nom parser for Versioning. Feel free to use this in combination with other general nom parsers.

pub fn is_ideal(&self) -> bool

A short-hand for detecting an inner SemVer.

pub fn is_general(&self) -> bool

A short-hand for detecting an inner Version.

pub fn is_complex(&self) -> bool

A short-hand for detecting an inner Mess.

pub fn nth(&self, n: usize) -> Option<u32>

Try to extract a position from the Versioning as a nice integer, as if it were a SemVer.

use versions::Versioning;

let semver = Versioning::new("1.2.3-r1+git123").unwrap();
assert!(semver.is_ideal());
assert_eq!(Some(1), semver.nth(0));
assert_eq!(Some(2), semver.nth(1));
assert_eq!(Some(3), semver.nth(2));

let version = Versioning::new("1:2.a.4.5.6.7-r1").unwrap();
assert!(version.is_general());
assert_eq!(Some(2), version.nth(0));
assert_eq!(None, version.nth(1));
assert_eq!(Some(4), version.nth(2));

let mess = Versioning::new("1.6a.0+2014+m872b87e73dfb-1").unwrap();
assert!(mess.is_complex());
assert_eq!(Some(1), mess.nth(0));
assert_eq!(None, mess.nth(1));
assert_eq!(Some(0), mess.nth(2));